+++ src/RunScript/RunScriptCommand.cs
@@ -63,18 +63,25 @@ namespace SCaddins.RunScript
             ElementSet elements)
         {
 #if !REVIT2026
-            SCaddinsApp.WindowManager.ShowMessageBox("Runscript is currently onl available in Revit 2026+");
-            return Result.Succeeded;
-#endif
-
+            SCaddinsApp.WindowManager.ShowMessageBox("Simple Mode Only");
             if (commandData == null)
             {
                 return Result.Failed;
             }
 
-            var vm = new ViewModels.RunScriptViewModel(commandData, elements);
-            SCaddinsApp.WindowManager.ShowDialogAsync(vm, null, ViewModels.RunScriptViewModel.DefaultViewSettings);
+            var vm = new ViewModels.RunScriptSimpleViewModel(commandData, elements);
+            SCaddinsApp.WindowManager.ShowDialogAsync(vm, null, ViewModels.RunScriptSimpleViewModel.DefaultViewSettings);
             return Result.Succeeded;
+#else
+            if (commandData == null)
+            {
+                return Result.Failed;
+            }
+
+            var vm2 = new ViewModels.RunScriptViewModel(commandData, elements);
+            SCaddinsApp.WindowManager.ShowDialogAsync(vm2, null, ViewModels.RunScriptViewModel.DefaultViewSettings);
+            return Result.Succeeded;
+#endif
         }
     }
 }