+++ src/RunScript/Views/RunScriptSimpleView.xaml.cs
@@ -1,0 +1,15 @@
+﻿using SCaddins.RunScript.ViewModels;
+using System;
+using System.Diagnostics;
+
+namespace SCaddins.RunScript.Views
+{
+    public partial class RunScriptSimpleView
+    {
+        public RunScriptSimpleView()
+        {
+            InitializeComponent();         
+        }
+    }
+}
+